combine multiple argo test files
